Split PyArray* and related function from lib.rs
